requirements.txt
flask==3.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==1.12.0